File_compressor_tool/requirements.txt
unittest2
pyhamcrest==2.0.4
pytest==7.4.0 
